+++ central/role/datastore/internal/store/store_impl.go
@@ -1,8 +1,7 @@
 package store
 
 import (
-	"fmt"
-
+	"github.com/pkg/errors"
 	rolePkg "github.com/stackrox/rox/central/role"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/bolthelper/crud/proto"
@@ -23,7 +22,7 @@ func (s *storeImpl) AddRole(role *storage.Role) error {
 // Pre-loaded roles cannot be updated./
 func (s *storeImpl) UpdateRole(role *storage.Role) error {
 	if isDefaultRole(role) {
-		return fmt.Errorf("cannot modify default role %s", role.GetName())
+		return errors.Errorf("cannot modify default role %s", role.GetName())
 	}
 	return s.roleCrud.Update(role)
 }
@@ -32,7 +31,7 @@ func (s *storeImpl) UpdateRole(role *storage.Role) error {
 // Pre-loaded roles cannot be removed.
 func (s *storeImpl) RemoveRole(name string) error {
 	if isDefaultRoleName(name) {
-		return fmt.Errorf("cannot modify default role %s", name)
+		return errors.Errorf("cannot modify default role %s", name)
 	}
 	return s.roleCrud.Delete(name)
 }