+++ migrator/runner/all.go
@@ -7,6 +7,7 @@ import (
 	_ "github.com/stackrox/rox/migrator/migrations/m_11_to_m_12_rewrite_alerts_and_deployments"
 	_ "github.com/stackrox/rox/migrator/migrations/m_12_to_m_13_default_retention_duration"
 	_ "github.com/stackrox/rox/migrator/migrations/m_13_to_m_14_flow_fixes_and_pruning"
+	_ "github.com/stackrox/rox/migrator/migrations/m_14_to_m_15_purge_debugmetrics"
 	_ "github.com/stackrox/rox/migrator/migrations/m_1_to_2_alert_violation"
 	_ "github.com/stackrox/rox/migrator/migrations/m_2_to_3_network_flows_in_badger"
 	_ "github.com/stackrox/rox/migrator/migrations/m_3_to_4_cluster_spec_to_status"